updated ft_lstat

+++ src/ft_list/ft_lstat.c
@@ -6,7 +6,7 @@
 /*   By: pducloux <pducloux@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/14 18:09:54 by pducloux          #+#    #+#             */
-/*   Updated: 2023/06/14 18:17:11 by pducloux         ###   ########.fr       */
+/*   Updated: 2023/06/15 17:21:18 by pducloux         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ t_list	*ft_lstat(t_list *lst, int index)
 	size_t	at;
 
 	i = ft_absi(index);
-	if (i >= ft_lstsize(lst))
+	if ((int)i >= ft_lstsize(lst))
 		return (NULL);
 	if (index < 0)
 		at = ft_lstsize(lst) + index;
